requirements.txt
# Azure Functions Runtime
azure-functions>=1.18.0

# Azure Services
azure-eventhub>=5.11.0
azure-data-tables>=12.4.0

# HTTP & API
requests>=2.31.0
urllib3>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0
